🚚 Rename: footer, headNav의 폴더를 새로 생성하여 집어넣음
