+++ pages/_app.tsx
@@ -3,7 +3,7 @@ import type { AppProps } from "next/app";
 import "@/styles/reset.css";
 import "@/styles/globals.css";
 import "@/styles/font-sizes.css";
-import Footer from "@/src/components/layout/Footer";
+import Footer from "@/src/components/layout/Footer/Footer";
 import { UserProvider } from "@/src/context/UserContext";
 
 export default function App({ Component, pageProps }: AppProps) {
