+++ pages/share.tsx
@@ -1,7 +1,7 @@
 import * as S from "@/styles/pages/share.style";
 
 // Components
-import HeadNav from "@/src/components/layout/HeadNav";
+import HeadNav from "@/src/components/layout/HeadNav/HeadNav";
 import ShareFolderProfile from "@/src/containers/share-page/profile/SharePageProfile";
 import LinkSearchBar from "@/src/components/link-card/link-search-form/LinkSearchBar";
 import LinkCardCollection from "@/src/components/link-card/link-card-collection/LinkCardCollection";
