+++ src/components/modal/remove-link/RemoveLink.tsx
@@ -8,7 +8,7 @@ import { ModalDataRemoveLink } from "@/src/types/ModalFunctionDataTypes";
  * @param {string} modalLink 현재 선택된 링크를 string으로 받아 삭제하는 기능을 가진 모달입니다.
  * @returns
  */
-export default function ModalRemoveLink({ modalData }: ModalDataRemoveLink) {
+export default function ModalRemoveLink({ modalData }: { modalData: string }) {
   return (
     <>
       <S.ModalTitle>
